@@ -13,10 +13,11 @@
     "build": "webpack --mode production"
   },
   "devDependencies": {
+    "@iconify/json": "^2.2.392",
     "@iconify/tailwind4": "^1.0.6",
-    "@tailwindcss/postcss": "^4.1.13",
+    "@tailwindcss/postcss": "^4.1.14",
     "@types/webpack": "^5.28.5",
-    "@vue/compiler-sfc": "^3.5.21",
+    "@vue/compiler-sfc": "^3.5.22",
     "copy-webpack-plugin": "^13.0.1",
     "css-loader": "^7.1.2",
     "css-minimizer-webpack-plugin": "^7.0.2",
@@ -24,20 +25,19 @@
     "html-webpack-plugin": "^5.6.4",
     "mini-css-extract-plugin": "^2.9.4",
     "ngrok": "^5.0.0-beta.2",
-    "npm-check-updates": "^18.1.1",
+    "npm-check-updates": "^19.0.0",
     "pinia": "^3.0.3",
     "postcss": "^8.5.6",
     "postcss-loader": "^8.2.0",
     "prettier": "^3.6.2",
     "prettier-plugin-tailwindcss": "^0.6.14",
-    "sass": "^1.92.1",
-    "sass-loader": "^16.0.5",
     "style-loader": "^4.0.0",
-    "tailwindcss": "^4.1.13",
-    "vue": "^3.5.21",
+    "tailwindcss": "^4.1.14",
+    "vanilla-tilt": "^1.8.1",
+    "vue": "^3.5.22",
     "vue-i18n": "^11.1.12",
     "vue-loader": "^17.4.2",
-    "webpack": "^5.101.3",
+    "webpack": "^5.102.0",
     "webpack-cli": "^6.0.1",
     "webpack-dev-server": "^5.2.2"
   }
